First Step Moving POI_Gen out of World_Gen
