+++ Runtime/Components/POI_Gen.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c0d82b9cfd6bfc84b90d4461ca856d13
+guid: ac9f50a623def2046afdfe76a7334a9e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
