--- Runtime/Components/POI_Gen_Adv.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac9f50a623def2046afdfe76a7334a9e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
